# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
  /orders/save:
    post:
      tags:
      - order-controller
      operationId: saveOrders
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Order'
        required: true
      responses:
        "200":
          description: OK
  /orders/get:
    get:
      tags:
      - order-controller
      operationId: getOrders
      parameters:
      - name: country
        in: query
        required: false
        schema:
          type: string
      - name: date
        in: query
        required: false
        schema:
          type: string
      - name: weightLimit
        in: query
        required: false
        schema:
          type: number
          format: double
      - name: pageNo
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /orders/getTotalWeight:
    get:
      tags:
      - order-controller
      operationId: getTotalWeightOfOrdersPerCountry
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeightPerCountry'
  /orders/getTotalOrders:
    get:
      tags:
      - order-controller
      operationId: getTotalOrdersPerCountry
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderPerCountry'
components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        phone_number:
          type: string
        parcel_weight:
          type: number
          format: double
        country:
          type: string
    WeightPerCountry:
      type: object
      properties:
        country:
          type: string
        totalWeight:
          type: number
          format: double
    OrderPerCountry:
      type: object
      properties:
        country:
          type: string
        count:
          type: integer
          format: int64